requirements.txt
crewai==1.10.1
psycopg2-binary
fastapi==0.109.0
uvicorn==0.27.0
tavily-python
apscheduler>=3.10.4
pytz
requests==2.31.0
groq
litellm
python-dotenv==1.0.0
Pillow
twilio==8.10.0
hubspot-api-client==8.2.0
anthropic==0.25.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
